--- Modules/FilterBottomSheet/DemoApp/DemoAppSources/ViewController.swift
@@ -1,16 +1,0 @@
-//
-//  ViewController.swift
-//  POLZZAK
-//
-//  Created by POLZZAK_iOS.
-//
-
-import UIKit
-
-final class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .white
-    }
-}
