--- Modules/FilterBottomSheet/DemoApp/DemoAppUITests/FilterBottomSheetDemoAppUITests.swift
@@ -1,24 +1,0 @@
-//
-//  FilterBottomSheetDemoAppUITests.swift
-//  POLZZAK
-//
-//  Created by POLZZAK_iOS.
-//
-
-import XCTest
-import Nimble
-
-final class FilterBottomSheetDemoAppUITests: XCTestCase {
-
-    var app: XCUIApplication!
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-        app = XCUIApplication()
-        app.launch()
-    }
-
-    func testExample() {
-        // UI tests code
-    }
-}
