--- Modules/FilterBottomSheet/Tests/FilterBottomSheetTests.swift
@@ -1,22 +1,0 @@
-//
-//  FilterBottomSheetModuleTests.swift
-//  POLZZAK
-//
-//  Created by POLZZAK_iOS.
-//
-
-import XCTest
-import Nimble
-@testable import FilterBottomSheet
-
-final class FilterBottomSheetTests: XCTestCase {
-
-    override func setUp() {
-        super.setUp()
-    }
-    override func tearDown() {
-        super.tearDown()
-    }
-    func testExample() {
-    }
-}
