+++ Modules/PolzzakUIKit/Resources/Assets.xcassets/filterButton.imageset/filterButton.png
