+++ Sources/DataLayer/Entities/StampBoard/MemberType.swift
