+++ Sources/DataLayer/Network/Enum/FilterType.swift
