+++ Sources/PresentationLayer/Common/ViewModelProtocol/TabFilterViewModelProtocol.swift
@@ -9,7 +9,6 @@ import Combine
 import Foundation
 
 import DynamicTabView
-import FilterBottomSheet
 
 protocol TabFilterViewModelProtocol: AnyObject {
     associatedtype DataListType
