+++ Sources/PresentationLayer/Common/ViewProtocol/CollectionConfigurable.swift
@@ -7,8 +7,6 @@
 
 import UIKit
 
-import FilterBottomSheet
-
 protocol CollectionLayoutConfigurable {
     var collectionView: UICollectionView! { get }
     func configureHeaderAndFooter(for section: NSCollectionLayoutSection, isDataNotEmpty: Bool, filterType: FilterType)
