+++ Sources/PresentationLayer/Coupon/CouponList/CouponListViewController.swift
@@ -11,7 +11,6 @@ import UIKit
 import CombineCocoa
 import DynamicTabView
 import Loading
-import FilterBottomSheet
 import SnapKit
 import PolzzakUIKit
 import PullToRefresh
@@ -283,11 +282,10 @@ extension CouponListViewController {
         case .section(let memberId):
             guard let section = viewModel.sectionOfMember(with: memberId) else { return }
             let family = viewModel.dataList.value[section].family
-            if viewModel.userType == .child {
-                filterView.handleChildSectionFilterButtonTap(with: family)
-            } else {
-                filterView.handleParentSectionFilterButtonTap(with: family)
-            }
+            let nickname = family.nickname
+            let memberName = family.memberType.detail ?? ""
+            let isParent = viewModel.userType == .parent
+            filterView.handleSectionFilterButtonTap(isParent: isParent, nickname: nickname, memberName: memberName)
         case .none:
             break
         }
