+++ Sources/PresentationLayer/Main/StampBoardList/MainViewModel.swift
@@ -9,7 +9,6 @@ import Combine
 import Foundation
 
 import DynamicTabView
-import FilterBottomSheet
 import Loading
 import PullToRefresh
